# api/openapi.yaml
openapi: 3.0.3
info:
  title: Task Management API
  version: 1.0.0
servers:
  - url: http://0.0.0.0:8080/api/v1
  - url: http://localhost:8080/api/v1
  - url: http://127.0.0.1:8080/api/v1
paths:
  /users:
    get:
      tags:
        - users
      summary: Get the current user's information
      operationId: getCurrentUser
      responses:
        "200":
          $ref: "#/components/responses/UserResponse"
        "400":
          $ref: "#/components/responses/ErrorResponse"
        "404":
          $ref: "#/components/responses/ErrorResponse"
      security:
        - CsrfAuth: []  # 認証が必須
    patch:
      tags:
        - users
      summary: Update the current user
      operationId: updateCurrentUser
      requestBody:
        $ref: "#/components/requestBodies/UserUpdateRequestBody"
        required: true
      responses:
        "200":
          $ref: "#/components/responses/UserResponse"
        "400":
          $ref: "#/components/responses/ErrorResponse"
        "404":
          $ref: "#/components/responses/ErrorResponse"
      security:
        - CsrfAuth: []  # 認証が必須
    delete:
      tags:
        - users
      summary: Delete the current user
      operationId: deleteCurrentUser
      responses:
        "204":
          description: User deleted
        "400":
          $ref: "#/components/responses/ErrorResponse"
        "404":
          $ref: "#/components/responses/ErrorResponse"
      security:
        - CsrfAuth: []  # 認証が必須
...
components:
  securitySchemes:
    CsrfAuth:
      type: apiKey
      in: header
      name: X-CSRF-TOKEN  # カスタムヘッダー名を指定
  schemas:
    UserRequest:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
          format: email
      required:
        - id
        - name
        - email  
...
    UserUpdateRequest:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
          format: email
        password:
          type: string
      required:
        - name
        - email
        - password
...
  requestBodies:
...
    UserUpdateRequestBody:
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/UserUpdateRequest"
...
  responses:            
    UserResponse:
      description: User response
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/UserRequest"
...
    ErrorResponse:
      description: Error response
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
            required:
              - message